+++ werk-runner/cache.rs
@@ -1,8 +1,11 @@
 use std::collections::BTreeMap;
 
 use stringleton::Symbol;
+use werk_eval::Io;
 use werk_fs::Absolute;
 use werk_util::hash128::Hash128;
+
+pub const WERK_CACHE_FILENAME: &str = ".werk-cache";
 
 /// The contents of `.werk-cache`.
 #[derive(Debug, Default, serde::Serialize, serde::Deserialize)]
@@ -10,6 +13,99 @@ pub struct WerkCache {
     /// Per-build-target caches.
     #[serde(default)]
     pub build: BTreeMap<Absolute<werk_fs::PathBuf>, TargetOutdatednessCache>,
+}
+
+impl WerkCache {
+    pub fn read(io: &dyn Io, output_dir: &Absolute<std::path::Path>) -> WerkCache {
+        let werk_cache_path = output_dir.join(WERK_CACHE_FILENAME).unwrap();
+        tracing::debug!("trying to read .werk-cache: {}", werk_cache_path.display());
+        let data = match io.read_file(&werk_cache_path) {
+            Ok(data) => data,
+            Err(err) => {
+                if err.kind() != std::io::ErrorKind::NotFound {
+                    tracing::error!("Failed to read workspace cache, even though it exists: {err}");
+                }
+                tracing::debug!(".werk-cache does not exist");
+                return WerkCache::default();
+            }
+        };
+
+        if data.is_empty() {
+            tracing::debug!(".werk-cache is empty");
+            return WerkCache::default();
+        }
+
+        match toml_edit::de::from_slice(&data) {
+            Ok(cache) => {
+                tracing::trace!(".werk-cache contents: {cache:#?}");
+                cache
+            }
+            Err(err) => {
+                tracing::error!("Failed to parse workspace cache: {err}");
+                WerkCache::default()
+            }
+        }
+    }
+
+    pub fn write(
+        &self,
+        io: &dyn Io,
+        output_dir: &Absolute<std::path::Path>,
+    ) -> std::io::Result<()> {
+        fn make_table(item: &mut toml_edit::Item) -> Option<&mut toml_edit::Table> {
+            match std::mem::take(item).into_table() {
+                Ok(table) => {
+                    *item = toml_edit::Item::Table(table);
+                    if let toml_edit::Item::Table(table) = item {
+                        Some(table)
+                    } else {
+                        unreachable!()
+                    }
+                }
+                Err(not_table) => {
+                    *item = not_table;
+                    None
+                }
+            }
+        }
+
+        let mut doc = match toml_edit::ser::to_document(self) {
+            Ok(data) => data,
+            Err(err) => {
+                tracing::error!("Serialization error writing .werk-cache: {err}");
+                panic!("Serialization error writing .werk-cache: {err}");
+            }
+        };
+
+        if let Some(build) = doc.get_mut("build") {
+            let build = make_table(build).expect("build is not a table");
+            build.set_implicit(true);
+            for (_target, target_info) in build.iter_mut() {
+                make_table(target_info);
+            }
+        }
+
+        let toml = format!("# Generated by werk. It can be safely deleted.\n\n{doc}");
+
+        let path = output_dir.join(WERK_CACHE_FILENAME).unwrap();
+        tracing::debug!("writing .werk-cache to {}", path.display());
+
+        if let Err(err) = io.create_parent_dirs(&path) {
+            tracing::error!(
+                "Error creating parent directory for .werk-cache '{}': {err}",
+                output_dir.display()
+            );
+            return Err(err);
+        }
+
+        match io.write_file(&path, toml.as_bytes()) {
+            Ok(()) => Ok(()),
+            Err(err) => {
+                tracing::error!("Error writing .werk-cache: {err}");
+                Err(err)
+            }
+        }
+    }
 }
 
 /// Per-target cache of used outdatedness information.
