+++ werk-runner/workspace.rs
@@ -13,9 +13,8 @@ use werk_parser::ast;
 use werk_util::{DiagnosticFileId, DiagnosticSpan, hash128::Hash128};
 
 use crate::{
-    Error, Render,
-    cache::{TargetOutdatednessCache, WerkCache},
-    ir,
+    Error, Manifest, Render,
+    cache::{TargetOutdatednessCache, WERK_CACHE_FILENAME, WerkCache},
 };
 
 #[derive(Clone)]
@@ -67,7 +66,7 @@ impl WorkspaceSettings {
 }
 
 pub struct Workspace {
-    pub manifest: ir::Manifest,
+    pub manifest: Manifest,
     // Project root - note that the workspace only accesses this directory
     // through the `Io` trait, and never directly.
     project_root: Absolute<std::path::PathBuf>,
@@ -98,8 +97,6 @@ struct Caches {
     build_recipe_hashes: HashMap<String, Hash128>,
 }
 
-pub const WERK_CACHE_FILENAME: &str = ".werk-cache";
-
 impl Workspace {
     pub fn new(
         io: Arc<dyn Io>,
@@ -107,7 +104,7 @@ impl Workspace {
         project_root: Absolute<std::path::PathBuf>,
         settings: &WorkspaceSettings,
     ) -> Result<Self, Error> {
-        let werk_cache = read_workspace_cache(&*io, &settings.output_directory);
+        let werk_cache = WerkCache::read(&*io, &settings.output_directory);
 
         let mut workspace_files =
             IndexMap::with_capacity_and_hasher(1024, ahash::RandomState::default());
@@ -136,7 +133,7 @@ impl Workspace {
         // results.
         workspace_files.sort_unstable_keys();
 
-        let manifest = ir::Manifest::default();
+        let manifest = Manifest::default();
 
         let workspace = Self {
             manifest,
@@ -458,7 +455,7 @@ impl Workspace {
     #[expect(clippy::unused_async)] // Preserving `async` for future-proofing.
     pub async fn finalize(&self) -> std::io::Result<()> {
         let cache = self.werk_cache.lock();
-        write_workspace_cache(&*self.io, &self.output_directory, &cache)
+        cache.write(&*self.io, &self.output_directory)
     }
 
     pub fn workspace_files(
@@ -674,97 +671,6 @@ pub(crate) fn compute_stable_semantic_hash<T: werk_util::SemanticHash + ?Sized>(
 
 fn compute_glob_hash(files: &[Absolute<werk_fs::PathBuf>]) -> Hash128 {
     compute_stable_hash(files)
-}
-
-fn read_workspace_cache(io: &dyn Io, output_dir: &Absolute<std::path::Path>) -> WerkCache {
-    let werk_cache_path = output_dir.join(WERK_CACHE_FILENAME).unwrap();
-    tracing::debug!("trying to read .werk-cache: {}", werk_cache_path.display());
-    let data = match io.read_file(&werk_cache_path) {
-        Ok(data) => data,
-        Err(err) => {
-            if err.kind() != std::io::ErrorKind::NotFound {
-                tracing::error!("Failed to read workspace cache, even though it exists: {err}");
-            }
-            tracing::debug!(".werk-cache does not exist");
-            return WerkCache::default();
-        }
-    };
-
-    if data.is_empty() {
-        tracing::debug!(".werk-cache is empty");
-        return WerkCache::default();
-    }
-
-    match toml_edit::de::from_slice(&data) {
-        Ok(cache) => {
-            tracing::trace!(".werk-cache contents: {cache:#?}");
-            cache
-        }
-        Err(err) => {
-            tracing::error!("Failed to parse workspace cache: {err}");
-            WerkCache::default()
-        }
-    }
-}
-
-fn write_workspace_cache(
-    io: &dyn Io,
-    output_dir: &Absolute<std::path::Path>,
-    cache: &WerkCache,
-) -> std::io::Result<()> {
-    fn make_table(item: &mut toml_edit::Item) -> Option<&mut toml_edit::Table> {
-        match std::mem::take(item).into_table() {
-            Ok(table) => {
-                *item = toml_edit::Item::Table(table);
-                if let toml_edit::Item::Table(table) = item {
-                    Some(table)
-                } else {
-                    unreachable!()
-                }
-            }
-            Err(not_table) => {
-                *item = not_table;
-                None
-            }
-        }
-    }
-
-    let mut doc = match toml_edit::ser::to_document(cache) {
-        Ok(data) => data,
-        Err(err) => {
-            tracing::error!("Serialization error writing .werk-cache: {err}");
-            panic!("Serialization error writing .werk-cache: {err}");
-        }
-    };
-
-    if let Some(build) = doc.get_mut("build") {
-        let build = make_table(build).expect("build is not a table");
-        build.set_implicit(true);
-        for (_target, target_info) in build.iter_mut() {
-            make_table(target_info);
-        }
-    }
-
-    let toml = format!("# Generated by werk. It can be safely deleted.\n\n{doc}");
-
-    let path = output_dir.join(WERK_CACHE_FILENAME).unwrap();
-    tracing::debug!("writing .werk-cache to {}", path.display());
-
-    if let Err(err) = io.create_parent_dirs(&path) {
-        tracing::error!(
-            "Error creating parent directory for .werk-cache '{}': {err}",
-            output_dir.display()
-        );
-        return Err(err);
-    }
-
-    match io.write_file(&path, toml.as_bytes()) {
-        Ok(()) => Ok(()),
-        Err(err) => {
-            tracing::error!("Error writing .werk-cache: {err}");
-            Err(err)
-        }
-    }
 }
 
 impl werk_util::DiagnosticSourceMap for &Workspace {
